+++ src/engine/Engine.ts
@@ -1,5 +1,5 @@
 import { EX_VERSION } from './';
-import './MonkeyPatch';
+import './Polyfill';
 import { ICanUpdate, ICanDraw, ICanInitialize } from './Interfaces/LifecycleEvents';
 import { ILoadable } from './Interfaces/ILoadable';
 import { Promise } from './Promises';
